universal profile
(issue #2426)
